requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
python-multipart==0.0.6
chromadb==0.4.15
numpy==1.26.4
python-dotenv==1.0.0
pydantic==2.5.0